+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 # LamaPlugin Changelog
 
 ## [Unreleased]
+### Added
+- Indent `esac`, `od`, `fi` on typing 
+
 ### Fixed
 - Documentation comment highlight
 - Completion/Go To Declaration for local symbols, which defined later than current